Refactoring: Deleted day of week, command.c to handle commands, added function comments
