+++ tp3-uart/main.h
@@ -14,17 +14,22 @@
 
 #include "serialPort.h"
 #include "rtc3231.h"
+#include "led.h"
+#include "command.h"
 #include <util/delay.h>
 #include <string.h>
 
 #define RX_BUFFER_SIZE 64
 
+uint8_t bcd_to_decimal(uint8_t bcd);
+uint8_t decimal_to_bcd(uint8_t decimal);
+
 extern volatile char rx_buffer[RX_BUFFER_SIZE];
 extern volatile uint8_t rx_index;
 extern volatile uint8_t message_ready;
 
-// LED control variables
-extern volatile uint8_t led_state;  // 0 = OFF, 1 = ON
+// Control variables
+extern volatile uint8_t mode;  // 0 = OFF, 1 = ON
 
 // State machine variables
 extern volatile uint8_t mode_change_requested;
